GE4-35 - Refactor OpenAI JSON schema types to shared types package
