+++ types/openai.go
@@ -2,6 +2,7 @@ package types
 
 import (
 	"NotificationManagement/models"
+	"encoding/json"
 )
 
 type OpenAIModelResponse struct {
@@ -25,4 +26,33 @@ func FromOpenAIModel(model *models.OpenAIModel) *OpenAIModelResponse {
 		CreatedAt: model.CreatedAt.Format(ResponseDateFormat),
 		UpdatedAt: model.UpdatedAt.Format(ResponseDateFormat),
 	}
+}
+
+// JSONSchemaProperty represents a property in a JSON schema
+type JSONSchemaProperty struct {
+	Type        string `json:"type"`
+	Description string `json:"description,omitempty"`
+}
+
+// JSONSchema represents a JSON schema structure
+type JSONSchema struct {
+	Type                 string                        `json:"type"`
+	Properties           map[string]JSONSchemaProperty `json:"properties"`
+	Required             []string                      `json:"required"`
+	AdditionalProperties bool                          `json:"additionalProperties"`
+}
+
+// MarshalJSON implements json.Marshaler interface
+func (j JSONSchema) MarshalJSON() ([]byte, error) {
+	return json.Marshal(struct {
+		Type                 string                        `json:"type"`
+		Properties           map[string]JSONSchemaProperty `json:"properties"`
+		Required             []string                      `json:"required"`
+		AdditionalProperties bool                          `json:"additionalProperties"`
+	}{
+		Type:                 j.Type,
+		Properties:           j.Properties,
+		Required:             j.Required,
+		AdditionalProperties: j.AdditionalProperties,
+	})
 }
